networking-crews/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "networking-crews"
version = "0.1.0"
description = "Networking AI Agent - CrewAI Platform crews for event discovery, booking, debrief, and coaching"
authors = ["Networking Agent Team"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = ">=0.86.0" }
httpx = ">=0.27.0"
psycopg2-binary = ">=2.9.9"
google-api-python-client = ">=2.100.0"
google-auth = ">=2.23.0"
pydantic = ">=2.0.0"

[tool.poetry.scripts]
networking_crews = "networking_crews.main:run"
train = "networking_crews.main:train"
